refactor(domain): split common.rs to satisfy 1000-line CI limit
Move client, pricing, validation, interactive recovery, errors, and
operation helpers into domain/common/ submodules while preserving the
existing common::* import surface for domain commands.
